python/src/gateway-service/requirements.txt
python-dotenv
Flask
pymongo
PyJWT
pika